Added static library VS solution to allow easy inclusion into larger projects. Updated FreeRTOS configuration macros.

+++ Chimera/sim/sim_chimera.hpp
@@ -18,11 +18,11 @@
 namespace ChimeraSim
 {
 
-  uint32_t millis();
+  size_t millis();
 
-  void delayMilliseconds(uint32_t val );
+  void delayMilliseconds( size_t val );
 
-  void delayMicroseconds(uint32_t val );
+  void delayMicroseconds( size_t val );
 
 }
